+++ src/components/package/show/components/package-settings/fields/index.js
@@ -3,3 +3,4 @@ export * from './automatically-select-titles';
 export * from './provider-token';
 export * from './package-token';
 export * from './custom-alternate-names';
+export * from './package-display-name';