+++ src/hooks/use-provider-packages/use-provider-packages.js
@@ -16,7 +16,7 @@ import {
 } from '../../constants';
 
 export const useProviderPackages = ({ providerId, searchParams }) => {
-  const namespace = useNamespace();
+  const [namespace] = useNamespace({ key: 'provider-packages' });
   const ky = useOkapiKy().extend({
     headers: {
       Accept: '*/*',
@@ -39,11 +39,11 @@ export const useProviderPackages = ({ providerId, searchParams }) => {
     count: PAGE_SIZE,
   };
 
-  const { data = {}, isFetching } = useQuery(
-    [namespace, paramsWithPageAndCount],
-    () => ky.get(`eholdings/providers/${providerId}/packages?${qs.stringify(paramsWithPageAndCount)}`).json(),
-    { keepPreviousData: true },
-  );
+  const { data = {}, isFetching } = useQuery({
+    queryKey: [namespace, paramsWithPageAndCount],
+    queryFn: () => ky.get(`eholdings/providers/${providerId}/packages?${qs.stringify(paramsWithPageAndCount)}`).json(),
+    keepPreviousData: true,
+  });
 
   return {
     isLoading: isFetching,