+++ src/routes/package-show-route/index.js
@@ -1,9 +1,6 @@
 import { connect } from 'react-redux';
 import { createResolver } from '../../redux';
 import { ProxyType } from '../../redux/application';
-import Package from '../../redux/package';
-import Provider from '../../redux/provider';
-import Resource from '../../redux/resource';
 import { selectPropFromData } from '../../redux/selectors';
 import {
   getAccessTypes as getAccessTypesAction,
@@ -19,18 +16,14 @@ import PackageShowRoute from './package-show-route';
 import { tagPaths } from '../../constants/tagPaths';
 
 export default connect(
-  (store, ownProps) => {
+  (store) => {
     const {
       eholdings: { data },
     } = store;
 
-    const { match } = ownProps;
     const resolver = createResolver(data);
-    const model = resolver.find('packages', match.params.packageId);
     return {
-      model,
       proxyTypes: resolver.query('proxyTypes'),
-      provider: resolver.find('providers', model.providerId),
       tagsModel: resolver.query('tags'),
       tagsModelOfAlreadyAddedTags: resolver.query('tags', undefined, { path: tagPaths.alreadyAddedToRecords }),
       resolver,
@@ -40,17 +33,11 @@ export default connect(
     };
   },
   {
-    getPackage: id => Package.find(id, { include: ['accessType'] }),
     getPackageTitles: getPackageTitlesAction,
     clearPackageTitles: clearPackageTitlesAction,
     getProxyTypes: () => ProxyType.query(),
     getTags: (params, options) => Tag.query(params, options),
-    getProvider: id => Provider.find(id),
-    unloadResources: collection => Resource.unload(collection),
-    updatePackage: model => Package.save(model),
     updateFolioTags: (model) => Tag.create(model),
-    destroyPackage: model => Package.destroy(model),
-    removeUpdateRequests: () => Package.removeRequests('update'),
     getAccessTypes: getAccessTypesAction,
     getCostPerUse: getCostPerUseAction,
     getCostPerUsePackageTitles: getCostPerUsePackageTitlesAction,