+++ src/routes/provider-show-route/provider-show-route.js
@@ -21,10 +21,11 @@ import {
   searchTypes,
   packageSortFilterConfig,
   selectionStatusFilterConfig,
+  contentTypeFilterConfig,
+  packageAccessFilterConfig,
 } from '../../constants';
 import { SearchSection } from '../../components/search-section';
 import PackageSearchFilters from '../../components/package-search-filters';
-import { contentTypeFilterConfig } from './package-search-filters-config';
 import { useProviderPackages } from '../../hooks';
 
 const propTypes = {
@@ -143,6 +144,7 @@ const ProviderShowRoute = ({
           packageSortFilterConfig,
           selectionStatusFilterConfig,
           contentTypeFilterConfig,
+          packageAccessFilterConfig,
         ]}
         {...props}
       />