+++ src/utils/serialize-provider.js
@@ -1,0 +1,7 @@
+export const serializeProviderAttributes = (values = {}) => {
+  return {
+    packagesSelected: values.packagesSelected || 0,
+    proxy: values.proxy || {},
+    providerToken: values.providerToken || {},
+  };
+};