+++ test/jest/__mock__/index.js
@@ -1,5 +1,6 @@
 import './currencyData.mock';
 import './reactIntl.mock';
+import './reactRouter.mock';
 import './stripesComponents.mock';
 import './stripesConfig.mock';
 import './stripesCore.mock';
@@ -7,3 +8,4 @@ import './stripesIcon.mock';
 import './stripesSmartComponent.mock';
 import './matchMedia.mock';
 import './resizeObserver.mock';
+import './hooks';